add event value -> tag copier
Add a new event processor that can copy a value from one event to
multiple other events.  If all tags present on the donor event are
identical on another event, apply the value as a new tag on the other
event.

The primary use case for this is when dealing with an output cache and
an Arista software device.  Arista splits all of its leaves into
separate Update messages, each with a different timestamp.  To avoid
losing precision of the individual counters, a merge is not desired, as
it would overwrite some timestamps, creating new TSDB datapoints where
they may not have existed before, and possibly throwing off delta
calculations.