Use K8s watch API in WatchServices
Instead of fetching the object on a schedule, we can ask Kubernetes to
send us changes as they happen. This is very similar to the Consul
method. The first response we get will include the ResourceVersion of
the current object, which we can use in subsequent requests to restart
the stream of object updates from where we left off.

This also changes the semantics of the watch timer to match the Consul
locker. Since we no longer poll for updates, this timer becomes a
timeout on the watch API call. This also adds a retry timer that is
used in case of an API error, again the same as the Consul locker.