Make the output/input fail when processor fails
In the case where the event processor fails to initialise
(could be for example a bad processor config, file not found,
starlark syntax error), the previous behaviour was to ignore
the processor.

However an event-processor can be crucial to the event processing
and ignoring it could lead to garbage in the downstream metrics
platform. It could lead for example to cardinality explosion.

This changes makes the input or output component fail if any
processor in the chain fails to initilise.