errChan in gnmi_server.go.Subscribe() is only used for case SubscriptionList_ONCE.
notification listener on a.c.Subscribe(sc.stream.Context(), ro) can't return on err
as this could block other sender on the same channel. Sender closes the channel
after sending the error and that would do.

in case of stream cancelation, error are popagated to gnmi_server.go.Subscribe()
for proper clean up.