Set minimal version for Go to 1.21.1
And set toolchain to 1.21.9. The "go" keyword in go.mod should be used
to specify the minimal version of Go required to compile the package.
The "toolchain" keyword is a suggestion of toolchain to use (and for the
main module, by default, it has the same effect as the "go" keyword).